[Refactor] 커스텀훅 분리 #1149
